requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
PyMySQL==1.1.0
cryptography==41.0.7
pycryptodome==3.19.0
python-dotenv==1.0.0
